Fix settings toolbar and wind tab showing only partial hours
- Settings: use setSupportActionBar + onOptionsItemSelected for reliable
  menu handling; explicit light tint on gear icon for dark theme visibility
- Wind: replace ScrollView with NestedScrollView so RecyclerView measures
  all 24 hourly items correctly instead of truncating
- Wind: change windDirection from Int to Double to match API response type
- Wind: fix formatDayLabel hardcoded "Dom" → use "Domani" for tomorrow

+++ MeteoRadar/app/src/main/java/it/shinyup/meteoradar/data/models/WeatherForecast.kt
@@ -56,7 +56,7 @@ data class WindForecastResponse(
 data class HourlyWindData(
     val time: List<String>,
     @SerializedName("wind_speed_10m") val windSpeed: List<Double>,
-    @SerializedName("wind_direction_10m") val windDirection: List<Int>,
+    @SerializedName("wind_direction_10m") val windDirection: List<Double>,
     @SerializedName("windgusts_10m") val windGusts: List<Double>,
     @SerializedName("temperature_2m") val temperature: List<Double>
 )
